Use Travis CI to run tests on all code changes
Every time someone submits a pull request, Travis Continuous Integration service can run the tests on it before it is reviewed.  Travis is always free to opensource projects like this one but the owner of the this repo would need to go to https://travis-ci.org/profile and flip the repository switch on to enable testing.